[#2] Split code into modular function
